// frontend/src/types/models.ts
export interface StudyPlan {
  plan_id: number;
  user: number; // FK → User.user_id
  title: string;
  start_date: string; // "YYYY-MM-DD" format
  end_date: string;   // "YYYY-MM-DD" format
}